requirements.txt
Adafruit-Blinka==6.10.3
adafruit-circuitpython-dht==3.6.1
Adafruit-PlatformDetect==3.14.2
Adafruit-PureIO==1.1.9
AWSIoTPythonSDK==1.4.9
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
paramiko==2.7.2
pycparser==2.20
pyftdi==0.53.1
PyNaCl==1.4.0
pyserial==3.5
pyusb==1.2.0
requests==2.25.1
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
six==1.16.0
sysv-ipc==1.1.0
urllib3==1.26.6
